Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3222



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@14113 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/tools/toolutil/uparse.c
@@ -385,7 +385,7 @@ u_parseUTF8(const char *source, int32_t sLen, char *dest, int32_t destCapacity, 
     int32_t i = 0;
     unsigned int value = 0;
     if(sLen == -1) {
-        sLen = strlen(source);
+        sLen = (int32_t)strlen(source);
     }
     
     while(read < source+sLen) {
